requirements.txt
Django==4.1.5
django-crispy-forms==1.14.0
python-dotenv==0.20.0
